clothing_labeling/requirements.txt
djangorestframework
django>=1.9.5